backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
greenlet>=3.0.3

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Vector Search
chromadb>=0.4.22
sentence-transformers>=2.3.1

# LLM Clients
openai>=1.60.0
ollama>=0.1.6

# Video & Subtitle Processing
# 复用 subgen 核心，或独立安装
faster-whisper>=1.1.1
whisper>=1.1.10

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0
tenacity>=8.2.3

# Task Queue (可选，用于后台处理)
celery>=5.3.6
redis>=5.0.1

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0
